Fixed 64bit issue. Adjusted for S2OSC

+++ osc/OscTypes.h
@@ -37,6 +37,7 @@
 #ifndef INCLUDED_OSCPACK_OSCTYPES_H
 #define INCLUDED_OSCPACK_OSCTYPES_H
 
+#include <cstdint>
 
 namespace osc{
 
@@ -63,13 +64,13 @@ typedef unsigned long long uint64;
 
 #if defined(__x86_64__) || defined(_M_X64)
 
-typedef signed int int32;
-typedef unsigned int uint32;
+typedef int32_t int32;
+typedef uint32_t uint32;
 
 #else
 
-typedef signed long int32;
-typedef unsigned long uint32;
+typedef int32_t int32;
+typedef uint32_t uint32;
 
 #endif
